# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateNoteRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UpdateNoteRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handler.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: '"success" atau "error"'
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: A simple and secure Notes API with user authentication and CRUD operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notes API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      summary: Register a new user
      tags:
      - Auth
  /health:
    get:
      description: Check API and database health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
  /notes:
    get:
      description: Get all notes for the authenticated user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get all notes
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Create a new note with title and content
      parameters:
      - description: Create note request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - Notes
  /notes/{id}:
    delete:
      description: Delete a note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - Notes
    get:
      description: Get a specific note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get a note by ID
      tags:
      - Notes
    put:
      consumes:
      - application/json
      description: Update a note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update note request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - Notes
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"